# src/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Haute Cuisine Documentation",
        "description": "APIs that manage the main functionalities for the haute cuisine web application"
    },
    "host": "localhost:8080/api/v1",
    "schemes": [
        "http",
        "https"
    ],
    "definitions": {
...
        "confirmOTPDTO": {
            "type": "object",
            "properties": {
                 "msisdn": {
                     "type": "string"
                 },
                 "code": {
                     "type": "string"
                 },
                 "countryCode": {
                     "type": "string"
                 }
            }
         },
...
         "resendDTO": {
            "type": "object",
            "properties": {
                 "msisdn": {
                     "type": "string"
                 },
                 "countryCode": {
                     "type": "string"
                 }
            }
         },
...
    },
...
    "paths": {
...
        "/users/confirm-otp":{
            "put": {
                "tags": [
                    "user"
                ],
                "summary": "route for users to confirm the otp",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "the user has to confirm the otp generated and sent",
                        "schema": {
                            "$ref": "#/definitions/confirmOTPDTO"
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "schema":{
                            "$ref": "#/definitions/success response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema":{
                            "$ref": "#/definitions/error response"
                        }
                    }
                },
                "operationId": "confirmOTP"
            }
        },
...
        "/users/resend":{
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "route for users to resend OTP (One Time Password)",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "the user has to confirm the otp generated and sent",
                        "schema": {
                            "$ref": "#/definitions/resendDTO"
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "schema":{
                            "$ref": "#/definitions/success response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema":{
                            "$ref": "#/definitions/error response"
                        }
                    }
                },
                "operationId": "resendOTP"
            }
        },
...
    }
}